+++ doc/Changelog.md
@@ -2,6 +2,10 @@
 
 # git master
 
+* [9](https://github.com/BlueBrain/Rockets/pull/9):
+  Added JSON-RPC 2.0 support over HTTP.
+* [8](https://github.com/BlueBrain/Rockets/pull/8):
+  Added JSON-RPC 2.0 parser with client-server over websockets.
 * [1](https://github.com/BlueBrain/Rockets/pull/1):
   Initial version.